packages/dara-core/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.blue]
line-length = 120

[tool.borg.scripts]
format = "poetry run isort dara && poetry run blue dara"
lint = "poetry run mypy --ignore-missing-imports -p dara.core && poetry run pylint dara"
package = "poetry build -f wheel"
publish = "poetry publish"
security-scan = "poetry run bandit dara -r -ll -i"
test = "DARA_TEST_FLAG=true poetry run pytest ./tests"

[tool.isort]
known_causalens = ["dara"]
profile = "black"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "CAUSALENS", "FIRSTPARTY", "LOCALFOLDER"]

[tool.poetry]
authors = ["Sam Smith <sam@causalens.com>", "Krzysztof Bielikowicz <krzysztof@causalens.com>"]
description = ""
include = ["dara/core/js_tooling/**/*", "dara/core/umd/**/*", "dara/core/log_configs/**/*"]
name = "dara-core"
version = "0.3.3"
license = "BUSL-1.1"

[[tool.poetry.packages]]
include = "dara"

[tool.poetry.dependencies]
async-asgi-testclient = "^1.4.11"
click = "8.1.3"
colorama = "^0.4.6"
croniter = "^1.0.15"
cryptography = ">=41.0.2"
fastapi = ">=0.95.1, <0.96.0"
fastapi-vite = "0.3.1"
httpx = ">=0.23.0"
jinja2 = ">=2.1.1, <3.1.0"
odfpy = "*"
openpyxl = "*"
packaging = "^23.1"
pandas = ">=1.1.0, <3.0.0"
prometheus-client = "^0.14.1"
pyarrow = "*"
pydantic = ">=1.8.1, <2.0.0"
pyjwt = ">=2.0.1, <3.0.0"
python = ">=3.8.0, <3.12.0"
python-dotenv = "^0.19.2"
python-multipart = "^0.0.5"
pyxlsb = "*"
responses = "^0.18.0"
tblib = "^1.7.0"
toml = "^0.10.2"
typing-extensions = ">=4.0.0, <4.6.0"
xlrd = "*"
dara-components = {version = "0.3.3", optional = true}
create-dara-app = {version = "0.3.3", optional = true}

[tool.poetry.extras]
all = ["dara-components", "create-dara-app"]

[tool.poetry.dependencies.uvicorn]
extras = ["standard"]
version = "^0.22.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.5"
blue = "^0.9.0"
isort = "^5.10.1"
mypy = "^0.991.0"
pylint = ">=2.6.0, <3.0.0"
pytest = "^7.0.0"
pytest-timeout = "^2.1.0"
requests = ">=2.25.1, <3.0.0"
types-croniter = "^1.0.10"
types-requests = "^2.27.29"
[tool.poetry.plugins."pydoc_markdown.interfaces.Renderer"]
docusaurus_internal = "docusaurus_internal:DocusaurusRenderer"

[tool.poetry.scripts]
dara = "dara.core.cli:cli"
[tool.pytest.ini_options]
timeout = 300
